@@ -1,6 +1,6 @@
 {
   "name": "angular-starter",
-  "version": "10.2.3",
+  "version": "11.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --port 4200",
@@ -15,7 +15,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/ganatan//angular10-app.git"
+    "url": "https://github.com/ganatan/angular11-app.git"
   },
   "author": {
     "name": "danny",
@@ -24,49 +24,48 @@
   },
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/ganatan/angular10-app/issues"
+    "url": "https://github.com/ganatan/angular11-app/issues"
   },
-  "homepage": "https://github.com/ganatan/angular10-app#readme",
+  "homepage": "https://github.com/ganatan/angular11-app#readme",
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.2.3",
-    "@angular/common": "10.2.3",
-    "@angular/compiler": "10.2.3",
-    "@angular/core": "10.2.3",
-    "@angular/forms": "10.2.3",
-    "@angular/platform-browser": "10.2.3",
-    "@angular/platform-browser-dynamic": "10.2.3",
-    "@angular/platform-server": "10.2.3",
-    "@angular/router": "10.2.3",
-    "angular-in-memory-web-api": "0.11.0",
-    "@angular/service-worker": "10.2.3",
+    "@angular/animations": "11.0.0",
+    "@angular/common": "11.0.0",
+    "@angular/compiler": "11.0.0",
+    "@angular/core": "11.0.0",
+    "@angular/forms": "11.0.0",
+    "@angular/platform-browser": "11.0.0",
+    "@angular/platform-browser-dynamic": "11.0.0",
+    "@angular/platform-server": "11.0.0",
+    "@angular/router": "11.0.0",
+    "@angular/service-worker": "11.0.0",
     "@asymmetrik/ngx-leaflet": "8.1.0",
     "@fortawesome/fontawesome-free": "5.15.1",
-    "@nguniversal/express-engine": "10.1.0",
-    "bootstrap": "5.0.0-alpha1",
+    "@nguniversal/express-engine": "11.0.0",
+    "bootstrap": "5.0.0-alpha3",
     "express": "4.17.1",
     "chart.js": "2.9.4",
     "leaflet": "1.7.1",
     "ng2-charts": "2.4.2",
+    "jquery": "3.5.1",
     "rxjs": "6.6.3",
     "tslib": "2.0.3",
     "zone.js": "0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1002.0",
-    "@angular/cli": "10.2.0",
-    "@angular/compiler-cli": "10.2.3",
-    "@nguniversal/builders": "10.1.0",
+    "@angular-devkit/build-angular": "0.1100.0",
+    "@angular/cli": "11.0.0",
+    "@angular/compiler-cli": "11.0.0",
+    "@nguniversal/builders": "11.0.0",
     "@types/express": "4.17.9",
-    "@types/node": "14.14.7",
     "@types/jasmine": "3.6.1",
-    "@types/jasminewd2": "2.0.8",
+    "@types/node": "14.14.7",
     "codelyzer": "6.0.1",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "6.0.0",
     "karma": "5.2.3",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-coverage": "2.0.3",
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.5.4",
     "protractor": "7.0.0",
