@@ -1,7 +1,6 @@
 import { TestBed } from '@angular/core/testing';
-
-import { ItemsService } from './items.service';
 import { HttpClientModule } from '@angular/common/http';
+import { ItemsService } from './items.service';
 
 describe('ItemsService', () => {
   let service: ItemsService;
@@ -10,7 +9,7 @@ describe('ItemsService', () => {
     TestBed.configureTestingModule({
       imports: [
         HttpClientModule
-      ],
+      ]
     });
     service = TestBed.inject(ItemsService);
   });
